+++ otp/example/totp_example.dart
@@ -1,4 +1,4 @@
-import 'package:another_authenticator_totp/totp_algorithm.dart';
+import 'package:another_authenticator_otp/otp/totp_algorithm.dart';
 
 void main() {
   print(Totp.generateCode(1542791843, "JBSWY3DPEHPK3PXP"));
