--- totp/lib/totp.dart
@@ -1,4 +1,0 @@
-library another_authenticator_totp;
-
-export 'models/otp_item.dart';
-export 'base32.dart';
